[geoclue-provider-hybris] Split adaptation code to backend. MER#1942
