+++ hal/hallocationbackend.pro
@@ -1,0 +1,9 @@
+include($$PWD/../geoclue-providers-hybris.pri)
+
+HEADERS += \
+    hallocationbackend.h
+
+SOURCES += \
+    hallocationbackend.cpp
+
+PKGCONFIG += libhardware android-headers
